requirements.txt
streamlit
pillow
pillow-heif
